requirements.txt
streamlit
fastapi
uvicorn
pillow
scikit-learn
pandas
numpy
matplotlib

